Parse push message in PushMessageParser
