+++ app/src/test/kotlin/at/bitfire/davdroid/push/PushMessageParserTest.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
+ */
+
+package at.bitfire.davdroid.push
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+class PushMessageParserTest {
+
+    private val parse = PushMessageParser()
+
+    @Test
+    fun testWithXmlDeclAndTopic() {
+        val topic = parse(
+            "<?xml version=\"1.0\" ?>" +
+            "<push-message xmlns='DAV:Push'>" +
+            "<topic>sample-topic</topic>" +
+            "</push-message>"
+        )
+        assertEquals("sample-topic", topic)
+    }
+
+}
